Moved configuration related code to common
