+++ pkg/common/common_suite_test.go
@@ -1,4 +1,4 @@
-package common_test
+package common
 
 import (
 	"testing"
