+++ pkg/common/config.go
@@ -19,12 +19,15 @@ package common
 import (
 	"encoding/json"
 	"errors"
+	"flag"
 	"fmt"
 	"os"
 	"strings"
 	"time"
 
+	"github.com/spf13/pflag"
 	"gopkg.in/yaml.v3"
+	"k8s.io/klog/v2"
 )
 
 const (
@@ -119,24 +122,24 @@ type LoraModule struct {
 }
 
 // Needed to parse values that contain multiple strings
-type MultiString struct {
-	Values []string
+type multiString struct {
+	values []string
 }
 
-func (l *MultiString) String() string {
-	return strings.Join(l.Values, " ")
+func (l *multiString) String() string {
+	return strings.Join(l.values, " ")
 }
 
-func (l *MultiString) Set(val string) error {
-	l.Values = append(l.Values, val)
+func (l *multiString) Set(val string) error {
+	l.values = append(l.values, val)
 	return nil
 }
 
-func (l *MultiString) Type() string {
+func (l *multiString) Type() string {
 	return "strings"
 }
 
-func (c *Configuration) UnmarshalLoras() error {
+func (c *Configuration) unmarshalLoras() error {
 	c.LoraModules = make([]LoraModule, 0)
 	for _, jsonStr := range c.LoraModulesString {
 		var lora LoraModule
@@ -165,7 +168,7 @@ func NewConfig() *Configuration {
 	}
 }
 
-func (c *Configuration) Load(configFile string) error {
+func (c *Configuration) load(configFile string) error {
 	configBytes, err := os.ReadFile(configFile)
 	if err != nil {
 		return fmt.Errorf("failed to read configuration file: %s", err)
@@ -175,10 +178,10 @@ func (c *Configuration) Load(configFile string) error {
 		return fmt.Errorf("failed to unmarshal configuration: %s", err)
 	}
 
-	return c.UnmarshalLoras()
+	return c.unmarshalLoras()
 }
 
-func (c *Configuration) Validate() error {
+func (c *Configuration) validate() error {
 	if c.Model == "" {
 		return errors.New("model parameter is empty")
 	}
@@ -264,4 +267,111 @@ func (c *Configuration) Validate() error {
 		return errors.New("ObjectToolCallNotRequiredParamProbability should be between 0 and 100")
 	}
 	return nil
+}
+
+// ParseCommandParamsAndLoadConfig loads configuration, parses command line parameters, merges the values
+// (command line values overwrite the config file ones), and validates the configuration
+func ParseCommandParamsAndLoadConfig() (*Configuration, error) {
+	config := NewConfig()
+
+	configFileValues := getParamValueFromArgs("config")
+	if len(configFileValues) == 1 {
+		if err := config.load(configFileValues[0]); err != nil {
+			return nil, err
+		}
+	}
+
+	servedModelNames := getParamValueFromArgs("served-model-name")
+	loraModuleNames := getParamValueFromArgs("lora-modules")
+
+	f := pflag.NewFlagSet("llm-d-inference-sim flags", pflag.ContinueOnError)
+
+	f.IntVar(&config.Port, "port", config.Port, "Port")
+	f.StringVar(&config.Model, "model", config.Model, "Currently 'loaded' model")
+	f.IntVar(&config.MaxNumSeqs, "max-num-seqs", config.MaxNumSeqs, "Maximum number of inference requests that could be processed at the same time (parameter to simulate requests waiting queue)")
+	f.IntVar(&config.MaxLoras, "max-loras", config.MaxLoras, "Maximum number of LoRAs in a single batch")
+	f.IntVar(&config.MaxCPULoras, "max-cpu-loras", config.MaxCPULoras, "Maximum number of LoRAs to store in CPU memory")
+	f.IntVar(&config.MaxModelLen, "max-model-len", config.MaxModelLen, "Model's context window, maximum number of tokens in a single request including input and output")
+
+	f.StringVar(&config.Mode, "mode", config.Mode, "Simulator mode, echo - returns the same text that was sent in the request, for chat completion returns the last message, random - returns random sentence from a bank of pre-defined sentences")
+	f.IntVar(&config.InterTokenLatency, "inter-token-latency", config.InterTokenLatency, "Time to generate one token (in milliseconds)")
+	f.IntVar(&config.TimeToFirstToken, "time-to-first-token", config.TimeToFirstToken, "Time to first token (in milliseconds)")
+	f.IntVar(&config.KVCacheTransferLatency, "kv-cache-transfer-latency", config.KVCacheTransferLatency, "Time for KV-cache transfer from a remote vLLM (in milliseconds)")
+	f.IntVar(&config.InterTokenLatencyStdDev, "inter-token-latency-std-dev", config.InterTokenLatencyStdDev, "Standard deviation for time between generated tokens (in milliseconds)")
+	f.IntVar(&config.TimeToFirstTokenStdDev, "time-to-first-token-std-dev", config.TimeToFirstTokenStdDev, "Standard deviation for time before the first token will be returned (in milliseconds)")
+	f.IntVar(&config.KVCacheTransferLatencyStdDev, "kv-cache-transfer-latency-std-dev", config.KVCacheTransferLatencyStdDev, "Standard deviation for time for KV-cache transfer from a remote vLLM (in milliseconds)")
+	f.Int64Var(&config.Seed, "seed", config.Seed, "Random seed for operations (if not set, current Unix time in nanoseconds is used)")
+
+	f.IntVar(&config.MaxToolCallIntegerParam, "max-tool-call-integer-param", config.MaxToolCallIntegerParam, "Maximum possible value of integer parameters in a tool call")
+	f.IntVar(&config.MinToolCallIntegerParam, "min-tool-call-integer-param", config.MinToolCallIntegerParam, "Minimum possible value of integer parameters in a tool call")
+	f.Float64Var(&config.MaxToolCallNumberParam, "max-tool-call-number-param", config.MaxToolCallNumberParam, "Maximum possible value of number (float) parameters in a tool call")
+	f.Float64Var(&config.MinToolCallNumberParam, "min-tool-call-number-param", config.MinToolCallNumberParam, "Minimum possible value of number (float) parameters in a tool call")
+	f.IntVar(&config.MaxToolCallArrayParamLength, "max-tool-call-array-param-length", config.MaxToolCallArrayParamLength, "Maximum possible length of array parameters in a tool call")
+	f.IntVar(&config.MinToolCallArrayParamLength, "min-tool-call-array-param-length", config.MinToolCallArrayParamLength, "Minimum possible length of array parameters in a tool call")
+	f.IntVar(&config.ToolCallNotRequiredParamProbability, "tool-call-not-required-param-probability", config.ToolCallNotRequiredParamProbability, "Probability to add a parameter, that is not required, in a tool call")
+	f.IntVar(&config.ObjectToolCallNotRequiredParamProbability, "object-tool-call-not-required-field-probability", config.ObjectToolCallNotRequiredParamProbability, "Probability to add a field, that is not required, in an object in a tool call")
+
+	// These values were manually parsed above in getParamValueFromArgs, we leave this in order to get these flags in --help
+	var dummyString string
+	f.StringVar(&dummyString, "config", "", "The path to a yaml configuration file. The command line values overwrite the configuration file values")
+	var dummyMultiString multiString
+	f.Var(&dummyMultiString, "served-model-name", "Model names exposed by the API (a list of space-separated strings)")
+	f.Var(&dummyMultiString, "lora-modules", "List of LoRA adapters (a list of space-separated JSON strings)")
+	// In order to allow empty arguments, we set a dummy NoOptDefVal for these flags
+	f.Lookup("served-model-name").NoOptDefVal = "dummy"
+	f.Lookup("lora-modules").NoOptDefVal = "dummy"
+
+	flagSet := flag.NewFlagSet("simFlagSet", flag.ExitOnError)
+	klog.InitFlags(flagSet)
+	f.AddGoFlagSet(flagSet)
+
+	if err := f.Parse(os.Args[1:]); err != nil {
+		if err == pflag.ErrHelp {
+			// --help - exit without printing an error message
+			os.Exit(0)
+		}
+		return nil, err
+	}
+
+	// Need to read in a variable to avoid merging the values with the config file ones
+	if loraModuleNames != nil {
+		config.LoraModulesString = loraModuleNames
+		if err := config.unmarshalLoras(); err != nil {
+			return nil, err
+		}
+	}
+	if servedModelNames != nil {
+		config.ServedModelNames = servedModelNames
+	}
+
+	if err := config.validate(); err != nil {
+		return nil, err
+	}
+
+	return config, nil
+}
+
+func getParamValueFromArgs(param string) []string {
+	var values []string
+	var readValues bool
+	for _, arg := range os.Args[1:] {
+		if readValues {
+			if strings.HasPrefix(arg, "--") {
+				break
+			}
+			if arg != "" {
+				values = append(values, arg)
+			}
+		} else {
+			if arg == "--"+param {
+				readValues = true
+				values = make([]string, 0)
+			} else if strings.HasPrefix(arg, "--"+param+"=") {
+				// Handle --param=value
+				values = append(values, strings.TrimPrefix(arg, "--"+param+"="))
+				break
+			}
+		}
+	}
+	return values
 }
